Add files via upload
2008_SAS_FW_2.130.404-4659 is the latest version, does not look like it is updating after that.
(This is the one currently on my LSI/BROADCOM card. MegaRAID SAS 9240-8i but supports the 4i also.)
(This card is made for server only but can 3d print holder for desktop. Price is much cheaper since not direct fit in desktop.)
(I think it supports 16 or 24 drives, but not sure)